@@ -1,6 +1,6 @@
 {
   "name": "CloudFlow",
-  "version": "0.1.4",
+  "version": "0.2.0",
   "license": "Apache-2.0",
   "scripts": {
     "ng": "ng",
@@ -15,15 +15,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "4.4.6",
-    "@angular/common": "4.4.6",
-    "@angular/compiler": "4.4.6",
-    "@angular/core": "4.4.6",
-    "@angular/forms": "4.4.6",
-    "@angular/http": "4.4.6",
-    "@angular/platform-browser": "4.4.6",
-    "@angular/platform-browser-dynamic": "4.4.6",
-    "@angular/router": "4.4.6",
+    "@angular/animations": "5.0.1",
+    "@angular/common": "5.0.1",
+    "@angular/compiler": "5.0.1",
+    "@angular/core": "5.0.1",
+    "@angular/forms": "5.0.1",
+    "@angular/http": "5.0.1",
+    "@angular/platform-browser": "5.0.1",
+    "@angular/platform-browser-dynamic": "5.0.1",
+    "@angular/router": "5.0.1",
     "@ng-bootstrap/ng-bootstrap": "1.0.0-beta.5",
     "angular-oauth2-oidc": "^2.1.3",
     "angular-split": "^0.2.2",
@@ -40,14 +40,14 @@
     "moment": "^2.18.1",
     "ngx-clipboard": "^8.0.4",
     "open-sans-fontface": "^1.4.0",
-    "rxjs": "^5.4.0",
+    "rxjs": "^5.5.0",
     "sweetalert2": "^6.11.1",
     "zone.js": "^0.8.16"
   },
   "devDependencies": {
-    "@angular/cli": "1.4.9",
-    "@angular/compiler-cli": "4.4.6",
-    "@angular/language-service": "4.4.6",
+    "@angular/cli": "1.5.0",
+    "@angular/compiler-cli": "5.0.1",
+    "@angular/language-service": "5.0.1",
     "@types/codemirror": "^0.0.43",
     "@types/dagre": "0.7.36",
     "@types/jasmine": "~2.5.53",
